#132 Ввод прочности в экипировку
