+++ Zilon.Core/Zilon.Core/Persons/SurvivalStat.cs
@@ -1,20 +1,17 @@
-﻿using System;
-using Zilon.Core.Common;
+﻿using Zilon.Core.Common;
 
 namespace Zilon.Core.Persons
 {
     /// <summary>
     /// Текущие показатели характеристики модуля выживания.
     /// </summary>
-    public class SurvivalStat
+    public sealed class SurvivalStat: Stat
     {
         private float _rawValue;
 
-        public SurvivalStat(int startValue, int min, int max)
+        public SurvivalStat(int startValue, int min, int max): base(startValue, min, max)
         {
-            Range = new Range<int>(min, max);
 
-            Value = startValue;
         }
 
         /// <summary>
@@ -22,43 +19,6 @@ namespace Zilon.Core.Persons
         /// </summary>
         public SurvivalStatType Type { get; set; }
 
-        /// <summary>
-        /// Текущее значение.
-        /// </summary>
-        public int Value
-        {
-            get
-            {
-                if (_rawValue == 1)
-                {
-                    return Range.Max;
-                }
-
-                if (_rawValue == 0)
-                {
-                    return Range.Min;
-                }
-
-                var result = Math.Round((Range.Max - Range.Min) * _rawValue + Range.Min);
-                return (int)result;
-            }
-            set
-            {
-                if (Range.Max == Range.Min)
-                {
-                    _rawValue = 1;
-                    return;
-                }
-
-                var boundedValue = Range.GetBounded(value);
-                _rawValue = (boundedValue - Range.Min) / (float)(Range.Max - Range.Min);
-            }
-        }
-
-        /// <summary>
-        /// Минимальное/максимальное значение.
-        /// </summary>
-        public Range<int> Range { get; private set; }
 
         /// <summary>
         /// Скорость снижения характеристики за ход.
@@ -67,22 +27,5 @@ namespace Zilon.Core.Persons
 
         /// <summary> Набор ключевых точек характеристики. </summary>
         public SurvivalStatKeyPoint[] KeyPoints { get; set; }
-
-        /// <summary>
-        /// Изменение текущего диапазона характеристики.
-        /// </summary>
-        /// <param name="min">The minimum.</param>
-        /// <param name="max">The maximum.</param>
-        public void ChangeStatRange(int min, int max)
-        {
-            if (min >= max)
-            {
-                Range = new Range<int>(min, min);
-                Value = Range.Min;
-                return;
-            }
-
-            Range = new Range<int>(min, max);
-        }
     }
 }
